+++ sdk/src/wallet/operations/transaction/mod.rs
@@ -90,7 +90,10 @@ where
         let block_id = match self
             .submit_basic_block(
                 Some(Payload::from(signed_transaction_data.payload.clone())),
-                options.as_ref().and_then(|options| options.issuer_id),
+                options
+                    .as_ref()
+                    .and_then(|options| options.issuer_id)
+                    .or(signed_transaction_data.issuer_id),
                 true,
             )
             .await