+++ bindings/nodejs/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 2.0.0-alpha.5 - 2024-04-15
+
+### Added
+
+- `{PreparedTransactionData, SignedTransactionData}::issuerId`;
+
+### Fixed
+
+- Implicit account transition with existing account;
+
 ## 2.0.0-alpha.4 - 2024-04-04
 
 ### Changed